added JWT claim verification
This is an improvement and a "bugfix", as this should be happening
alright, but I only figured this out now.

Both supported JWT libs only verify the cryptographic signature, and do
not verify claims out-of-the-box, as this is a user-level concern.
+ruby-jwt+ provides it as a few kwarg options, where one can pass what
to verify, and in some cases how; for `json-jwt`, we're on our own.

jti verification was also a big concern, as it was using the signing key
as part of its generation, which can't be used in separate AS/RS
scenarios. (It was relaced by an hmac secret, which is also not perfect
as it has to be shared). More work to be done in this department.

jwt_decode() can also be called with `verify: false` now, which is great
if we don't need to verify it, which we'll need.