making use of id_token_signed_response_alg option to select the signing key
this was done by reshaping the way keys are declared in the auth server,
in that now, there's a new option, `oauth_jwt_keys`, indexing algos to a
key (or group of keys). This then makes it pick it up when encoding the
id token. This new option will supersede the legacy token options.